% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export}
\alias{export}
\title{Export results from set- and network-based enrichment analysis}
\usage{
export(res, to = c("enrichplot", "GOPlot"))
}
\arguments{
\item{res}{Enrichment analysis results as returned by \code{\link{sbea}}
or \code{\link{nbea}}.}

\item{to}{Character. Downstream package to which export enrichment analysis
results to. Defaults to \code{"enrichplot"}, currently the only supported
export option.}
}
\description{
This function exports results of differential expression (DE) analysis such
as enrichplot and GOPlot.
}
