% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMicrobeSets.R
\name{getMicrobeSets}
\alias{getMicrobeSets}
\title{Definition of microbe sets according to different sources}
\usage{
getMicrobeSets(
  db = c("bugsigdb", "manalyst", "mpattern", "mdirectory"),
  tax.id.type = c("ncbi", "metaphlan"),
  cache = TRUE,
  ...
)
}
\arguments{
\item{db}{Database from which microbe sets should be retrieved. Currently, 
either 'bugsigdb' (default), 'manalyst', 'mpattern', or 'mdirectory'.}

\item{tax.id.type}{Character. Taxonomic ID type of the returned microbe sets.
Currently either 'ncbi' (default) or 'metaphlan'.}

\item{cache}{Logical.  Should a locally cached version used if available?
Defaults to \code{TRUE}.}

\item{...}{Additional arguments for individual microbe set databases.
For \code{db = "manalyst"}: \itemize{ \item lib: Character. MicrobiomeAnalyst
taxon set library. Options include taxon sets associated with human genetic
variations ('gene'), host-intrinsic ('host_int'), host-extrinsic ('host_ext'),
environmental ('env'), and microbiome-intrinsic ('mic_int') factors.
See references.}}
}
\description{
Functionality for retrieving microbe sets from databases such as
BugSigDB and the MicrobeDirectory.
}
\references{
BugSigDB: \url{https://bugsigdb.org}

MicrobiomeAnalyst: \url{https://www.microbiomeanalyst.ca}
}
