% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{vst}
\alias{vst}
\title{Variance-stabilizing transformation for RNA-seq expression data}
\usage{
vst(se)
}
\arguments{
\item{se}{An object of class \code{\linkS4class{SummarizedExperiment}}.}
}
\value{
An object of class \code{\linkS4class{SummarizedExperiment}}.
}
\description{
This function implements a variance-stabilizing transformation (VST) for 
RNA-seq read count data. It accounts for differences in sequencing depth
between samples and over-dispersion of read count data. Permutation-based
enrichment methods can then be applied as for microarray data.
}
\details{
The VST uses the cpm function implemented in the edgeR package to compute 
moderated log2 read counts. Using edgeR's estimate of the common dispersion 
phi, the prior.count parameter of the cpm function is chosen as 0.5 / phi as 
previously suggested (Harrison, 2015).
}
\examples{

    se <- makeExampleData(what="SE", type="rseq")
    vstSE <- vst(se) 

}
\references{
Harrison (2015) Anscombe's 1948 variance stabilizing transformation for
the negative binomial distribution is well suited to RNA-seq expression
data. doi:10.7490/f1000research.1110757.1

Anscombe (1948) The transformation of Poisson, binomial and
negative-binomial data. Biometrika 35(3-4):246-54.

Law et al. (2014) voom: precision weights unlock linear model analysis tools 
for RNA-seq read counts. Genome Biol 15:29.
}
\seealso{
\code{\link{cpm}} and \code{\link{estimateDisp}}
}
\author{
Ludwig Geistlinger <Ludwig.Geistlinger@sph.cuny.edu>
}
