% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deAna.R
\name{deAna}
\alias{deAna}
\alias{de.ana}
\title{Differential expression analysis between two sample groups}
\usage{
deAna(expr, grp = NULL, blk = NULL, de.method = c("limma", "edgeR",
  "DESeq2"), padj.method = "BH", stat.only = FALSE,
  filter.by.expr = TRUE)
}
\arguments{
\item{expr}{Expression data.  A numeric matrix. Rows correspond to genes,
columns to samples.  Alternatively, this can also be an object of class
\code{\linkS4class{SummarizedExperiment}}.}

\item{grp}{*BINARY* group assignment for the samples.  Use '0' and '1' for
unaffected (controls) and affected (cases) samples, respectively.  If NULL,
this is assumed to be defined via a column named 'GROUP' in the
\code{\link{colData}} slot if 'expr' is a
\code{\linkS4class{SummarizedExperiment}}.}

\item{blk}{Optional. For paired samples or sample blocks.  This can also be
defined via a column named 'BLOCK' in the \code{\link{colData}} slot if
'expr' is a \code{\linkS4class{SummarizedExperiment}}.}

\item{de.method}{Differential expression method.  Use 'limma' for microarray
and RNA-seq data.  Alternatively, differential expression for RNA-seq data
can be also calculated using edgeR ('edgeR') or DESeq2 ('DESeq2').  Defaults
to 'limma'.}

\item{padj.method}{Method for adjusting p-values to multiple testing.  For
available methods see the man page of the stats function
\code{\link{p.adjust}}.  Defaults to 'BH'.}

\item{stat.only}{Logical. Should only the test statistic be returned?  This
is mainly for internal use, in order to carry out permutation tests on the
DE statistic for each gene.  Defaults to FALSE.}

\item{filter.by.expr}{Logical. For RNA-seq data: include only genes with
sufficiently large counts in the DE analysis? If TRUE, excludes genes not 
satisfying a minimum number of read counts across samples using the 
\code{\link{filterByExpr}} function from the edgeR package.
Defaults to TRUE.}
}
\value{
A DE-table with measures of differential expression for each
gene/row, i.e. a two-column matrix with log2 fold changes in the 1st column
and derived p-values in the 2nd column.  If 'expr' is a
\code{\linkS4class{SummarizedExperiment}}, the DE-table will be
automatically appended to the \code{\link{rowData}} slot.
}
\description{
The function carries out a differential expression analysis between two
sample groups. Resulting fold changes and derived p-values are returned.
Raw p-values are corrected for multiple testing.
}
\examples{

    # (1) microarray data: intensity measurements
    maSE <- makeExampleData(what="SE", type="ma")
    maSE <- deAna(maSE)
    rowData(maSE, use.names=TRUE)
    
    # (2) RNA-seq data: read counts
    rseqSE <- makeExampleData(what="SE", type="rseq")
    rseqSE <- deAna(rseqSE, de.method="DESeq2")
    rowData(rseqSE, use.names=TRUE)

}
\seealso{
\code{\link{readSE}} for reading expression data from file,
\code{\link{normalize}} for normalization of expression data,
\code{\link{voom}} for preprocessing of RNA-seq data, \code{\link{p.adjust}}
for multiple testing correction, \code{\link{eBayes}} for DE analysis with
limma, \code{\link{glmFit}} for DE analysis with edgeR, and
\code{DESeq} for DE analysis with DESeq2.
}
\author{
Ludwig Geistlinger <Ludwig.Geistlinger@sph.cuny.edu>
}
