% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIds.R
\name{map.ids}
\alias{map.ids}
\alias{id.types}
\title{Mapping between gene ID types for the rownames of a SummarizedExperiment}
\usage{
map.ids(se, org = NA, from = "ENSEMBL", to = "ENTREZID")

id.types(org)
}
\arguments{
\item{se}{An object of class \code{\linkS4class{SummarizedExperiment}}.
Expects the names to be of gene ID type given in argument 'from'.}

\item{org}{Organism in KEGG three letter code, e.g. \sQuote{hsa} for
\sQuote{Homo sapiens}.  See references.}

\item{from}{Gene ID type from which should be mapped.  Corresponds to the
gene ID type of the featureNames of argument 'se'.  Defaults to 'ENSEMBL'.}

\item{to}{Gene ID type to which should be mapped.  Corresponds to the gene
ID type the featuresNames of argument 'se' should be updated with.  Defaults
to 'ENTREZID'.}
}
\value{
id.types: character vector listing the available gene ID types for
the mapping;

map.ids: An object of \code{\linkS4class{SummarizedExperiment}}.
}
\description{
Functionality to map the rownames of a SummarizedExperiment between common
gene ID types such as ENSEMBL and ENTREZ.
}
\details{
The function 'id.types' lists the valid values which the arguments 'from'
and 'to' can take. This corresponds to the names of the available gene ID
types for the mapping.
}
\examples{

    # create an expression dataset with 3 genes and 3 samples
    se <- makeExampleData("SE", nfeat=3, nsmpl=3)
    names(se) <- paste0("ENSG00000000", c("003","005", "419"))
    se <- map.ids(se, org="hsa")

}
\references{
KEGG Organism code
\url{http://www.genome.jp/kegg/catalog/org_list.html}
}
\seealso{
\code{\linkS4class{SummarizedExperiment}}, \code{\link{mapIds}},
\code{\link{keytypes}}
}
\author{
Ludwig Geistlinger <Ludwig.Geistlinger@sph.cuny.edu>
}
